requirement.txt
alabaster==0.7.9
Babel==2.3.4
colorama==0.3.7
colored-traceback==0.2.2
colorlog==2.10.0
Django==1.10.4
django-admin-view-permission==0.6
django-annoying==0.10.3
django-extensions==1.7.7
django-logentry-admin==1.0.2
django-redis==4.7.0
djangorestframework==3.5.3
docutils==0.13.1
imagesize==0.7.1
invoke==0.14.0
Jinja2==2.8.1
MarkupSafe==0.23
psycopg2==2.6.2
Pygments==2.1.3
pygraphviz==1.3.1
pytz==2016.10
redis==2.10.5
requests==2.12.4
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.1
Werkzeug==0.11.15
